+++ apps/meteor/client/views/admin/settings/EditableSettingsProvider.tsx
@@ -8,7 +8,6 @@ import type { FilterOperators } from 'mongodb';
 import type { MutableRefObject, ReactNode } from 'react';
 import React, { useEffect, useMemo, useRef } from 'react';
 
-import { useLicenseBase } from '../../../hooks/useLicense';
 import { createReactiveSubscriptionFactory } from '../../../lib/createReactiveSubscriptionFactory';
 import type { EditableSetting, EditableSettingsContextValue } from '../EditableSettingsContext';
 import { EditableSettingsContext } from '../EditableSettingsContext';
@@ -193,13 +192,6 @@ const EditableSettingsProvider = ({ children, query = defaultQuery, omit = defau
 		Tracker.flush();
 	});
 
-	const { data } = useLicenseBase({
-		select: (data) => ({ isEnterprise: Boolean(data?.license.license), activeModules: data?.license.activeModules }),
-	});
-
-	const isEnterprise = data?.isEnterprise ?? false;
-	const activeModules = useMemo(() => data?.activeModules ?? [], [data]);
-
 	const contextValue = useMemo<EditableSettingsContextValue>(
 		() => ({
 			queryEditableSetting,
@@ -207,10 +199,8 @@ const EditableSettingsProvider = ({ children, query = defaultQuery, omit = defau
 			queryGroupSections,
 			queryGroupTabs,
 			dispatch,
-			isEnterprise,
-			activeModules,
 		}),
-		[queryEditableSetting, queryEditableSettings, queryGroupSections, queryGroupTabs, dispatch, isEnterprise, activeModules],
+		[queryEditableSetting, queryEditableSettings, queryGroupSections, queryGroupTabs, dispatch],
 	);
 
 	return <EditableSettingsContext.Provider children={children} value={contextValue} />;
